refactor(monorepo): Phase 8 — split env/config + remove Tauri from shared
Two structural changes:

1. **@openloomi/ui-runtime** — new UI-side package that statically imports
   the tauri runtime APIs. Owns platform/env.ts (isClient, isTauri,
   isBrowser, getPlatformKind), platform/filesystem.ts, and the
   tauri/browser filesystem adapters. Lives in the ui/ sub-project after
   Phase 9.

2. **@openloomi/env-config** — new dep-free env / deployment-mode /
   Tauri-path constants package. Holds the three pure leaves moved out of
   apps/web/lib/env/ (client-constants, client-mode, tauri-paths). The
   server-only constants.ts and server-constants.ts stay in apps/web
   for now — they pull in @/lib/db/utils (DUMMY_PASSWORD) and will move
   in a later phase once the DB layer is daemonized.

After this phase:
- @openloomi/shared no longer statically depends on the tauri runtime
  (verified by `git grep -l @tauri-apps packages/shared/src`).
- shared re-exports platform/* from ui-runtime for backward compat.
- apps/web/lib/env/{client-constants,client-mode,tauri-paths}.ts become
  re-export shims over @openloomi/env-config/{subpath}.

Build + typecheck + 2253 tests pass.
