+++ packages/env-config/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@openloomi/shared",
+  "name": "@openloomi/env-config",
   "version": "0.9.0",
   "type": "module",
   "files": ["dist", "README.md", "LICENSE"],
@@ -11,30 +11,22 @@
   "types": "./dist/index.d.ts",
   "exports": {
     ".": "./dist/index.js",
-    "./errors": "./dist/errors.js"
+    "./client-constants": "./dist/client-constants.js",
+    "./client-mode": "./dist/client-mode.js",
+    "./tauri-paths": "./dist/tauri-paths.js"
   },
   "repository": {
     "type": "git",
     "url": "https://github.com/melandlabs/openloomi.git",
-    "directory": "packages/shared"
+    "directory": "packages/env-config"
   },
-  "homepage": "https://github.com/melandlabs/openloomi/tree/main/packages/shared",
+  "homepage": "https://github.com/melandlabs/openloomi/tree/main/packages/env-config",
   "bugs": {
     "url": "https://github.com/melandlabs/openloomi/issues"
   },
   "license": "Apache-2.0",
-  "description": "Shared utility primitives used across OpenLoomi packages.",
-  "keywords": ["openloomi", "shared", "utilities"],
-  "dependencies": {
-    "@openloomi/ui-runtime": "workspace:*",
-    "clsx": "^2.1.1",
-    "tailwind-merge": "^2.5.2",
-    "zod": "^4.3.6",
-    "date-fns": "^4.1.0"
-  },
-  "peerDependencies": {
-    "react": ">=18.0.0"
-  },
+  "description": "Environment, deployment-mode and Tauri path constants used by both UI and runtime services.",
+  "keywords": ["openloomi", "env", "config", "tauri"],
   "devDependencies": {
     "@openloomi/config": "workspace:*",
     "typescript": "^5.6.3"
