+++ packages/shared/src/platform/index.ts
@@ -1,9 +1,16 @@
+// Phase 8: Platform detection moved to @openloomi/ui-runtime.
+// shared no longer statically depends on the tauri runtime.
 export {
   getFileSystem,
   type PlatformFileSystem,
   type SaveFileOptions,
   type DirEntry,
   type ListDirectoryOptions,
-} from "./filesystem";
+} from "@openloomi/ui-runtime/platform/filesystem";
 
-export { isClient, isTauri, isBrowser, getPlatformKind } from "./env";
+export {
+  isClient,
+  isTauri,
+  isBrowser,
+  getPlatformKind,
+} from "@openloomi/ui-runtime/platform/env";
