+++ packages/ui-runtime/package.json
@@ -1,0 +1,44 @@
+{
+  "name": "@openloomi/ui-runtime",
+  "version": "0.9.0",
+  "type": "module",
+  "files": ["dist", "README.md", "LICENSE"],
+  "publishConfig": {
+    "access": "public"
+  },
+  "main": "./dist/index.js",
+  "module": "./dist/index.js",
+  "types": "./dist/index.d.ts",
+  "exports": {
+    ".": "./dist/index.js",
+    "./platform/env": "./dist/platform/env.js",
+    "./platform/filesystem": "./dist/platform/filesystem.js",
+    "./platform/adapters/tauri/filesystem": "./dist/platform/adapters/tauri/filesystem.js",
+    "./platform/adapters/browser/filesystem": "./dist/platform/adapters/browser/filesystem.js"
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/melandlabs/openloomi.git",
+    "directory": "packages/ui-runtime"
+  },
+  "homepage": "https://github.com/melandlabs/openloomi/tree/main/packages/ui-runtime",
+  "bugs": {
+    "url": "https://github.com/melandlabs/openloomi/issues"
+  },
+  "license": "Apache-2.0",
+  "description": "UI-side runtime primitives (Tauri platform detection, browser/Tauri filesystem adapters).",
+  "keywords": ["openloomi", "ui", "tauri", "platform"],
+  "dependencies": {
+    "@tauri-apps/api": "^2.11.0",
+    "@tauri-apps/plugin-dialog": "^2.7.1",
+    "@tauri-apps/plugin-fs": "^2.5.1"
+  },
+  "devDependencies": {
+    "@openloomi/config": "workspace:*",
+    "typescript": "^5.6.3"
+  },
+  "scripts": {
+    "build": "tsup",
+    "prepublishOnly": "pnpm build"
+  }
+}
