+++ packages/ui-runtime/src/platform/index.ts
@@ -4,4 +4,6 @@ export {
   type SaveFileOptions,
   type DirEntry,
   type ListDirectoryOptions,
-} from "@openloomi/ui-runtime/platform/filesystem";
+} from "./filesystem";
+
+export { isClient, isTauri, isBrowser, getPlatformKind } from "./env";
